http - bigquery/requirements.txt
google-cloud-bigquery==3.30.0
functions-framework==3.*


